requirements.txt
Flask==3.1.0
psycopg2-binary
requests==2.31.0
